FULL_finance_platform/Finance_platform/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Validation & Serialization
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Excel Processing
openpyxl==3.1.2
xlsxwriter==3.1.9

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1
pytesseract==0.3.10
pillow==10.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.1

# HTTP & API
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
rich==13.7.0

# Caching & Background Tasks
redis==5.0.1
celery==5.3.4

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# AWS/Cloud (optional)
boto3==1.34.10

# AI/ML (optional)
anthropic>=0.39.0
